Add support for SRV records for finding IRC servers. eg irc://quakenet.org/
Of the top 20 networks, it looks like only quakenet support it right now,
but it makes sense and allows for multiple servers to exist on the same IP
with different ports without needing to know the ports.

Eg for a test network you could just connect to irc://inspircd.testnetwork/
or irc://ircu.testnetwork/ without needing to remember the ports.

Change-Id: Ia366292e64b2ca46cb8658b34fdac598fa155722