Remove chain of Comparators in output queue.
Instead of having lots of objects implement Comparable or Comparator,
pass the comparator in to the PriorityBlockingQueue.

This requires recreating the queue if the QueueHandler is ever changed,
but the previous behaviour resulted in an unstable/undefined ordering,
so it's probably better this way anyway.

This means that QueueItems no longer need to know about QueueHandler,
which makes me slightly happier.